chore : AttendanceResponse 에서만 사용하는 AttendanceInfo class를 inner class로 변경(#36)
